Add files via upload
The aim of this project is to develop an expert system that assists users in determining potential health conditions based on their symptoms. The system will ask users a series of questions regarding their symptoms and medical history to suggest possible ailments and recommend whether to consult a healthcare professional.